prepare for macos support
 - adjust CMakeLists for macos support
 - adjust Makefile for macos support
 - add missing headers
 - add option to configure cmake prefix path for Makefile build
 - fix C++ and LLVM debug compilation on macos for backends.mlir test
 - set mlir to static linkage (otherwise segfaults)
 - update dependency arrow to version 20
 - fix tpc* generate scripts for macos
 - add switch for asm parser architecture (x86 / arm)

+++ src/runtime/storage/LingoDBTable.cpp
@@ -205,7 +205,7 @@ void access(std::vector<size_t> colIds, lingodb::runtime::RecordBatchInfo* info,
       colInfo = chunk.getColumnInfo(colId);
       colInfo.offset += offset;
    }
-   info->numRows = std::min(currChunk->num_rows() - offset, numRows);
+   info->numRows = std::min(static_cast<size_t>(currChunk->num_rows() - offset), numRows);
 }
 
 } // namespace
